+++ CHANGELOG.md
@@ -45,6 +45,7 @@ const defaultLayoutPluginInstance = defaultLayoutPlugin({
 
 **Bug fixes**
 
+-   Don't highlight the entire page when selecting multiple lines
 -   The highlight plugin throws an exception when double click a page without selecting any text
 -   The default layout plugin throws an exception if the `setInitialTabFromPageMode` function returns a `Promise` which resolves an invalid tab index